server/requirements.txt
django>=5.1
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-decouple>=3.8